Change to float types

+++ custom_components/zaptec/misc.py
@@ -91,13 +91,13 @@ def mc_nbfx_decoder(msg: bytes) -> None:
             raise AttributeError(f"Unknown record type {hex(record_type)}")
 
 
-def get_ocmf_max_reader_value(data: dict) -> int:
+def get_ocmf_max_reader_value(data: dict) -> float:
     """Return the maximum reader value from OCMF data."""
 
     if not isinstance(data, dict):
-        return 0
+        return 0.
     return max(
-        int(reading.get("RV", 0))
+        float(reading.get("RV", 0.))
         for reading in data.get("RD",[])
     )
